CHROMIUM: topology: remove dmic16k from max98357a-waves-2way
After kernel backporting commit chain landed on v5.10 and DSP
migrated to rpl-001 for Brya (ADL-P), DSP panic issue is found on
Taniks, max98357a-rt56882-waves-2way topology.

By investigation DSP panic issue is still reproducible even if
Waves modules are removed. In the meanwhile, it will be not
reproducible by applying either of the following modifications:
1. Move dmic48k to Core#0 as well as dmic16k
2. Remove dmic16k pipeline and GHD pipeline

The second fix won't affect the behavior for Taniks because GHD is
not activated on Taniks in practice. It should be fair as the quick
fix. From my perspective the root cause might be related to
multi-core processings, sof-git issue is filed for futher tracking.

Signed-off-by: Pin-chih Lin <johnylin@google.com>
Change-Id: Idee86d1d617d2c68e1bd8b4f20c3d014d77e2091
Reviewed-on: https://chrome-internal-review.googlesource.com/c/chromeos/third_party/sound-open-firmware-private/+/5325553
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>